2021-2-15 12-1.

+++ Sources/SwiftXlsxwriter/Workbook.swift
@@ -35,7 +35,7 @@ public class Workbook {
         return Format(fmt)
     }
     
-    public func close() -> Int {
+    public func close() -> UInt8 {
         return workbook_close(wb)
     }
 }
